drm: move the card open and master to the device-context
The DRM card open + master acquisition moves from the per-view DrmBackend up to
DrmDisplay (the device-context), so a card's single master is shared by every
view that scans out to it rather than each view grabbing its own. DrmDisplay::
SharedDevice acquires the card lazily on first use -- via the libseat session
when present, else a direct open + drmSetMaster behind the foreground-VT guard,
which moves here too -- and holds it. The EGL DRM backend takes a non-owning
pointer to it through Create and no longer opens the device or takes master;
DrmDisplay drops master and closes the fd on its own teardown, after the
backends that used it are gone.

The DRM-KMS-Vulkan backend is unchanged: MakeDrmVulkanBackend does not call
SharedDevice, so a Vulkan view still opens its own device and self-masters
exactly as before. Mixing the two on one card is not supported.

Behavior-neutral for single-view: EGL renders 60fps with master now logged from
DrmDisplay, and Vulkan renders unchanged (self-mastered).
