+++ shell/display/drm_display.cc
@@ -17,8 +17,19 @@
 #include "display/drm_display.h"
 #include "logging/logging.h"
 
+#include <cerrno>
 #include <cstring>
+#include <optional>
 #include <utility>
+
+#include <fcntl.h>
+#include <linux/vt.h>
+#include <sys/ioctl.h>
+#include <sys/sysmacros.h>
+#include <unistd.h>
+#include <xf86drm.h>
+
+#include <drm-cxx/core/device.hpp>
 
 #include "backend/drm_kms_egl/drm_session.h"
 #include "cursor_kind.h"
@@ -55,6 +66,97 @@ drm::input::InputDeviceOpener OpenerFrom(homescreen::DrmSession* session) {
   return session->InputOpener();
 }
 
+// Refuse to run unless our controlling terminal is the *foreground* VT on
+// the kernel console. Rationale: drmSetMaster can succeed from a non-
+// foreground VT (e.g. a terminal emulator, SSH, or an inactive text VT)
+// when no compositor currently holds master, but the GPU keeps scanning
+// out whatever the foreground VT owns. Atomic commits get silently
+// accepted and no PAGE_FLIP_EVENT ever fires, so PresentLayers stalls on
+// the next flip wait — the "master-acquired but black screen" pathology
+// we kept hitting. Fail fast with an actionable message instead.
+//
+// Returns true if it's safe to proceed, false with a logged error if not.
+// A missing /dev/tty0 (container / headless service) is not fatal —
+// those cases are outside the scope of this check.
+bool VerifyForegroundVt(const std::string& drm_device) {
+  const int tty0 = ::open("/dev/tty0", O_RDONLY | O_CLOEXEC);
+  if (tty0 < 0) {
+    spdlog::warn(
+        "[DrmDisplay] open(/dev/tty0): {} — skipping foreground-VT check",
+        std::strerror(errno));
+    return true;
+  }
+  struct vt_stat vtstat{};
+  const bool got_state = ::ioctl(tty0, VT_GETSTATE, &vtstat) == 0;
+  const int vt_errno = errno;
+  ::close(tty0);
+  if (!got_state) {
+    spdlog::warn("[DrmDisplay] VT_GETSTATE: {} — skipping foreground-VT check",
+                 std::strerror(vt_errno));
+    return true;
+  }
+  const unsigned int active_vt = vtstat.v_active;
+
+  // open("/dev/tty") redirects to the process's controlling terminal, so
+  // calling open() against the special inode is the only way to acquire
+  // an fd that *targets* the ctty rather than the /dev/tty special device
+  // itself. ENXIO from open() means the process has no ctty at all (e.g.
+  // a daemon without setsid + TIOCSCTTY).
+  const int ctl_fd = ::open("/dev/tty", O_RDONLY | O_CLOEXEC | O_NOCTTY);
+  if (ctl_fd < 0) {
+    spdlog::error(
+        "[DrmDisplay] no controlling terminal (open /dev/tty: {}). Cannot "
+        "drive {} without a foreground VT — switch to the active console "
+        "(tty{}) and rerun.",
+        std::strerror(errno), drm_device, active_vt);
+    return false;
+  }
+  // fstat() of the resulting fd does NOT return the underlying tty's
+  // device numbers — it returns /dev/tty's own (5, 0) inode metadata on
+  // every Linux kernel from 4.x onward. TIOCGDEV is the canonical ioctl
+  // for retrieving the actual tty device number from a /dev/tty fd;
+  // available since Linux 2.6.18.
+  unsigned int dev = 0;
+  const int ioctl_rc = ::ioctl(ctl_fd, TIOCGDEV, &dev);
+  const int ioctl_errno = errno;
+  ::close(ctl_fd);
+  if (ioctl_rc != 0) {
+    spdlog::error(
+        "[DrmDisplay] TIOCGDEV(/dev/tty): {}. Cannot drive {} without "
+        "a foreground VT — switch to the active console (tty{}) and rerun.",
+        std::strerror(ioctl_errno), drm_device, active_vt);
+    return false;
+  }
+
+  // TTY_MAJOR is 4 on Linux; /dev/ttyN lives at (4, N). Terminal emulators
+  // and SSH sessions give /dev/tty a pts major (136+), which is precisely
+  // the case we want to refuse.
+  constexpr unsigned int kTtyMajor = 4;
+  const unsigned int ctl_major = major(static_cast<dev_t>(dev));
+  const unsigned int ctl_minor = minor(static_cast<dev_t>(dev));
+  if (ctl_major != kTtyMajor) {
+    spdlog::error(
+        "[DrmDisplay] controlling terminal is not a kernel VT "
+        "(major={}, minor={}) — you're running from a terminal emulator or "
+        "SSH. Active VT is tty{}. Switch to tty{} (Ctrl+Alt+F{}) and rerun, "
+        "or `sudo systemctl isolate multi-user.target` first. Refusing to "
+        "drive {}.",
+        ctl_major, ctl_minor, active_vt, active_vt, active_vt, drm_device);
+    return false;
+  }
+  if (ctl_minor != active_vt) {
+    spdlog::error(
+        "[DrmDisplay] controlling VT is tty{} but foreground VT is tty{} — "
+        "scanout goes to the foreground. Switch to tty{} (Ctrl+Alt+F{}) and "
+        "rerun. Refusing to drive {}.",
+        ctl_minor, active_vt, ctl_minor, ctl_minor, drm_device);
+    return false;
+  }
+  spdlog::info("[DrmDisplay] foreground VT check: on tty{} (active)",
+               active_vt);
+  return true;
+}
+
 }  // namespace
 
 DrmDisplay::DrmDisplay(int32_t width,
@@ -66,6 +168,7 @@ DrmDisplay::DrmDisplay(int32_t width,
       height_(height),
       refresh_rate_hz_(refresh_rate_hz),
       device_path_(std::move(device_path)),
+      no_seat_(no_seat),
       output_provider_(device_path_),
       session_(OpenSessionOrLog(no_seat)),
       seat_(std::make_unique<homescreen::DrmSeat>(width,
@@ -100,7 +203,80 @@ DrmDisplay::DrmDisplay(int32_t width,
   }
 }
 
-DrmDisplay::~DrmDisplay() = default;
+DrmDisplay::~DrmDisplay() {
+  // Drop master (and let drm_dev_'s destructor close the fd) once the backends
+  // that scanned out through it are gone. This display is shared and destroyed
+  // after them, so master is released last -- whoever takes over next (fbcon,
+  // a getty, the next compositor) inherits a clean device.
+  if (drm_master_ && drm_dev_.has_value()) {
+    drmDropMaster(drm_dev_->fd());
+    drm_master_ = false;
+  }
+}
+
+drm::Device* DrmDisplay::SharedDevice() {
+  if (drm_dev_.has_value()) {
+    return &*drm_dev_;
+  }
+
+  if (session_ != nullptr) {
+    // libseat path: the seat provider (logind/seatd/builtin) owns VT
+    // activation and master handoff; TakeDevice returns a master-capable fd
+    // while we hold the seat. No drmSetMaster, no foreground-VT guard.
+    const int fd = session_->TakeDevice(device_path_);
+    if (fd < 0) {
+      spdlog::error("[DrmDisplay] session take_device({}): failed",
+                    device_path_);
+      return nullptr;
+    }
+    drm_dev_.emplace(drm::Device::from_fd(fd));
+    spdlog::info("[DrmDisplay] opened {} via libseat (fd={})", device_path_,
+                 fd);
+  } else {
+    // Fallback path: no seat backend (or --drm-no-seat). Refuse up-front if we
+    // are not on the active VT, then open + take master directly.
+    if (no_seat_) {
+      spdlog::warn(
+          "[DrmDisplay] --drm-no-seat: skipping the foreground-VT guard on "
+          "{}. If the screen stays black, another DRM master holds the device "
+          "or scanout is owned by a different VT.",
+          device_path_);
+    } else if (!VerifyForegroundVt(device_path_)) {
+      return nullptr;
+    }
+
+    auto dev = drm::Device::open(device_path_);
+    if (!dev) {
+      spdlog::error("[DrmDisplay] open({}): {}", device_path_,
+                    dev.error().message());
+      return nullptr;
+    }
+    drm_dev_.emplace(std::move(*dev));
+
+    // Atomic writes from a non-master fd are silently accepted by some drivers
+    // (amdgpu) as no-ops, which presents as a blank panel with no
+    // PAGE_FLIP_EVENT. Refuse to run without master.
+    if (drmSetMaster(drm_dev_->fd()) != 0) {
+      if (const int err = errno;
+          err == EBUSY || err == EACCES || err == EPERM) {
+        spdlog::error(
+            "[DrmDisplay] cannot acquire DRM master on {} ({}). Another "
+            "display server (gdm / gnome-shell / sddm / Xorg / a Wayland "
+            "compositor) is already holding the device. Stop it or run from a "
+            "bare TTY (e.g. `sudo systemctl isolate multi-user.target`).",
+            device_path_, std::strerror(err));
+      } else {
+        spdlog::error("[DrmDisplay] drmSetMaster({}): {}", device_path_,
+                      std::strerror(err));
+      }
+      drm_dev_.reset();
+      return nullptr;
+    }
+    drm_master_ = true;
+    spdlog::info("[DrmDisplay] DRM master on {}", device_path_);
+  }
+  return &*drm_dev_;
+}
 
 void DrmDisplay::StartEvents() {
   if (seat_) {
